backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.7.1
asttokens==3.0.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docx==0.2.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==2.2.1
fastapi==0.118.0
fastjsonschema==2.21.2
filelock==3.19.1
fsspec==2025.9.0
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.35.3
ibm-db==3.2.1
idna==3.4
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.3.1
Markdown==3.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
murmurhash==1.0.13
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numpy==2.3.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.5.0
preshed==3.0.10
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.21
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.7.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2023.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.31.0
rich==14.1.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.15.2
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.16.0
smart_open==7.3.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
trio==0.23.1
trio-websocket==0.11.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.37.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
whitenoise==6.11.0
wrapt==1.17.3
wsproto==1.2.0
yarg==0.1.9
